src/adk/requirements.txt
numpy==1.24.4

google-auth
google-auth-oauthlib
google-api-python-client
google-cloud-aiplatform
google-adk
google-cloud-pubsub
google-cloud-secret-manager

python-dotenv
requests

slack_sdk
slack_bolt

google-cloud-firestore
google-cloud-bigquery